Move NoScript icon into URL bar. This fixes draggability dead area that was present
Fixes https://github.com/brave/browser-laptop/issues/5792

Auditors: @bradleyrichter, @luixxiul

Includes: Fix noscript tests
Auditors: @diracdeltas
