+++ less/navigationBar.less
@@ -572,20 +572,6 @@
   }
 }
 
-.noScriptButtonContainer {
-    display: flex;
-    margin-left: 6px;
-
-    .noScript {
-      font-size: 16px;
-      color: @braveOrange;
-
-      &:hover {
-        color: @braveOrange;
-      }
-  }
-}
-
 #navigator {
   .stopButton,
   .reloadButton,
